requirements.txt
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
